feature: добавляет блок выбора времени для сессии
