+++ src/components/SessionPlanner/SessionPlanner.css
@@ -39,7 +39,7 @@
   height: 48px;
   color: black;
   border: 1px solid;
-  border-color: var(--border-color-dark-violet);
+  border-color: var(--text-color-dark-blue);
   border-radius: 12px;
   font-family: var(--font-family-lato);
   font-size: 18px;
@@ -57,6 +57,11 @@
   color: var(--text-color-dark-blue);
   font-size: 18px;
   line-height: 1.33;
+}
+
+.session-planner_style_open {
+  color: var(--text-color-dark-violet);
+  border-color: var(--text-color-dark-violet);
 }
 
 .session-planner__button-add {
@@ -78,8 +83,8 @@
 
 .session-planner__timing-list {
   list-style: none;
-  /* display: none; */
-  display: flex;
+  display: none;
+  /* display: flex; */
   flex-direction: column;
   position: absolute;
   top: 114.5%;
@@ -103,11 +108,11 @@
 }
 
 .session-planner__timing-item {
+  /* display: none; */
   font-family: var(--font-family-lato);
   font-size: 18px;
   line-height: 1.33;
   font-weight: var(--text-weight-primary);
-  display: none;
   border: none;
   padding: 6px 0;
   background-color: inherit;
