+++ src/components/SessionPlanner/SessionPlanner.jsx
@@ -1,36 +1,18 @@
-import React, { useState } from 'react';
+import React from 'react';
 import './SessionPlanner.css';
 import Button from '../generic/Button/Button';
 import { TIMING_HOURS, TIMING_MINUTES } from '../../constants/db';
 import TimePicker from '../generic/TimePicker/TimePicker';
 
 export default function SessionPlanner() {
-  const [isShowDropdown, setIsShowDropdown] = useState(false);
-
-  const handleButtonClick = (e) => {
-    e.preventDefault();
-    if (!isShowDropdown) {
-      setIsShowDropdown(true);
-    } else {
-      setIsShowDropdown(false);
-    }
-  };
   return (
     <div className="session-planner">
       <div className="session-planner__time-picker">
         <h2 className="session-planner__text">Время начала сессии</h2>
         <div className="session-planner__timing-box">
-          <TimePicker
-            onClick={handleButtonClick}
-            timingList={TIMING_HOURS}
-            isShow={isShowDropdown}
-          />
+          <TimePicker id={1} timingList={TIMING_HOURS} />
           <p className="session-planner__time-colon">:</p>
-          <TimePicker
-            onClick={handleButtonClick}
-            timingList={TIMING_MINUTES}
-            isShow={isShowDropdown}
-          />
+          <TimePicker id={2} timingList={TIMING_MINUTES} />
         </div>
       </div>
       <Button type="submit" className="session-planner__button-add">
