+++ src/components/generic/TimePicker/TimePicker.jsx
@@ -4,27 +4,40 @@ import PropTypes from 'prop-types';
 import vectorOpened from './vector_opened.svg';
 import vectorClosed from './vector_closed.svg';
 
-export default function TimePicker({ onClick, timingList, isShow }) {
-  const classesItems = `session-planner__timing-item  ${
-    isShow ? 'session-planner__item_visible' : ''
+export default function TimePicker({ timingList }) {
+  const [isShowDropdown, setIsShowDropdown] = useState(false);
+  const classesItems = `session-planner__timing-list  ${
+    isShowDropdown ? 'session-planner__list_visible' : ''
   }`;
   const [isValueTime, setIsValueTime] = useState('');
 
+  const classOpen = `${isShowDropdown ? 'session-planner_style_open' : ''}`;
+
+  const handleButtonClick = (e) => {
+    e.preventDefault();
+    if (!isShowDropdown) {
+      setIsShowDropdown(true);
+    } else {
+      setIsShowDropdown(false);
+    }
+  };
+
   function changeSelect(e) {
     setIsValueTime(e.target.value);
+    setIsShowDropdown(false);
   }
 
   const titleValue = isValueTime || timingList[0];
 
   return (
-    <div className="session-planner__select">
-      <p className="session-planner__number">{titleValue}</p>
+    <div className={`session-planner__select ${classOpen}`}>
+      <p className={`session-planner__number ${classOpen}`}>{titleValue}</p>
       <button
         className="session-planner__button-drop"
-        type="button"
-        onClick={onClick}
+        type="submit"
+        onClick={handleButtonClick}
       >
-        {!isShow ? (
+        {!isShowDropdown ? (
           <img
             className="session-planner__vector-icon"
             src={vectorClosed}
@@ -38,13 +51,13 @@ export default function TimePicker({ onClick, timingList, isShow }) {
           />
         )}
       </button>
-      <div className="session-planner__timing-list">
+      <div className={classesItems}>
         {timingList.map((el) => (
           <input
             onClick={changeSelect}
             type="button"
             value={el}
-            className={classesItems}
+            className="session-planner__timing-item"
             key={el}
           />
         ))}
@@ -54,11 +67,5 @@ export default function TimePicker({ onClick, timingList, isShow }) {
 }
 
 TimePicker.propTypes = {
-  onClick: PropTypes.func,
   timingList: PropTypes.arrayOf(PropTypes.string).isRequired,
-  isShow: PropTypes.bool.isRequired,
-};
-
-TimePicker.defaultProps = {
-  onClick: () => {},
 };
